pyproject.toml
[project]
name = "demeter-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "colorama>=0.4.6",
    "fastapi[standard,standart]>=0.119.0",
    "httpx>=0.28.1",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.3.0",
    "python-json-logger>=4.0.0",
    "sqlalchemy>=2.0.44",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.37.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "flake8-annotations>=3.2.0",
    "flake8-bugbear>=24.12.12",
    "flake8-builtins>=3.0.0",
    "flake8-comprehensions>=3.17.0",
    "flake8-docstrings>=1.7.0",
    "flake8-import-order>=0.19.2",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pylint>=4.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.0",
]
